审计整改(AUDIT_2026Q3): 前端体验增强与后端资金正确性/启动安全
- 前端 T19: 分类页类目内榜单+面包屑(复用ProductLeaderboard, 按category_id拉取)

- 前端 T20: 购物车角标加购 bump 动效(主/次/移动端三处)

- 前端 T22: 购物车满减进度条+凑单入口(消费preview.full_reduce_progress)

- 前端 C-4: 生产构建 esbuild.drop 清除 console/debugger

- 后端 T16: affiliate/invoice/shipping 金额字段 Float-,2) + 迁移0013

- 后端 T18: _ensure_demo_columns 受 ALLOW_SCHEMA_AUTOFIX 开关控制(默认False, 生产安全)

- 后端: 修复迁移0011/0012 缺失的 op 导入与_COLUMNS 4元组解包错误

- CI: 新增 schema-gate 步骤(纯Alembic建库+ALLOW_SCHEMA_AUTOFIX=False验证)

+++ backend/app/models/affiliate.py
@@ -3,7 +3,7 @@ import enum
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, DateTime, Enum as SAEnum, Float, ForeignKey, Integer, String
+from sqlalchemy import Column, DateTime, Enum as SAEnum, ForeignKey, Integer, Numeric, String
 
 from app.db.base import Base
 
@@ -55,8 +55,8 @@ class AffiliateCommission(Base):
     order_id = Column(String(36), ForeignKey("orders.id"), nullable=False, unique=True)
     promoter_id = Column(String(36), ForeignKey("users.id"), nullable=False, index=True)
     buyer_id = Column(String(36), ForeignKey("users.id"), nullable=False)
-    order_amount = Column(Float, nullable=False)
-    commission = Column(Float, nullable=False)
+    order_amount = Column(Numeric(12, 2), nullable=False)
+    commission = Column(Numeric(12, 2), nullable=False)
     status = Column(SAEnum(CommissionStatus), default=CommissionStatus.SETTLED, nullable=False)
     created_at = Column(DateTime(timezone=True), default=_now)
 
@@ -74,7 +74,7 @@ class AffiliateWithdrawal(Base):
 
     id = Column(String(36), primary_key=True, default=_uuid)
     user_id = Column(String(36), ForeignKey("users.id"), nullable=False, index=True)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     status = Column(SAEnum(WithdrawalStatus), default=WithdrawalStatus.PENDING, nullable=False)
     remark = Column(String(200), nullable=True)
     created_at = Column(DateTime(timezone=True), default=_now)
