+++ backend/app/models/shipping.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, String
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Numeric, String
 from sqlalchemy.orm import relationship
 
 from app.db.base import Base
@@ -18,8 +18,8 @@ class ShippingTemplate(Base):
     id = Column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     merchant_id = Column(String(36), ForeignKey("users.id"), nullable=False, index=True)
     name = Column(String(120), nullable=False)
-    base_fee = Column(Float, default=0.0, nullable=False)
-    free_amount = Column(Float, default=0.0, nullable=False)
+    base_fee = Column(Numeric(12, 2), default=0, nullable=False)
+    free_amount = Column(Numeric(12, 2), default=0, nullable=False)
     is_default = Column(Boolean, default=False, nullable=False)
     created_at = Column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
